binder: remove unnecessary err check
In function IPCThreadState::transact(), data.errorCheck() will
be executed twice. Since IPCThreadState::transact() is the
critical path for binder call, it is better to do data.errorCheck()
once.

This patch removes the first check at the beginning of
IPCThreadState::transact(), the effect of this change is that LOG_ONEWAY(...)
will be executed in error case("data.errorCheck() != NO_ERROR")

I think this is not a problem. As "data.errorCheck() == NO_ERROR" is the
normal case(which will execute LOG_ONEWAY()), and even in error case,
we print a log about src/dst pid is not a bad idea.

Change-Id: I7b892a2294774c55ce0df56edee6a820f82c6f13
Signed-off-by: Ganesh Mahendran <opensource.ganesh@gmail.com>