libbinder_ndk: make AIBinder_isRemote const.
It's a C function, so it is not const, but it now takes a const
AIBinder object.

There isn't a way to do a check on a const IBinder if it is remote. We
could store the isRemote status as a member of AIBinder to avoid this,
but this would be wasted space. A const isRemote method can't be added
to IBinder for legacy reasons.

Bug: 111445392
Test: ./ndk/runtests.sh
Change-Id: I72df61e429e06e24dcffd4158447670ec1b41d7a