binderDriverInterfaceTest: fix failure
After servicemanager was re-written in C++, it started returning a
4-byte payload for ping. Whereas before, in the C implementation, a zero
byte payload was returned.

Allowing multiple values in this test to stop presubmit failures. As a
follow-up, I will try to determine if we can switch libbinder to a
0-byte ping (and revert this CL). Alternatively, once SM is in all
branches, if we want this behavior, we can restrict this test to 4-bytes
OR we can remove this expected value from this test altogether.

This test is allowing both values and being fixed in this strange way
because an investigation will follow, and our branching structure (and
vendor prebuilts) make it hard to revert the offending change.

Fixes: 136633463
Bug: 137203409
Test: binderDriverInterfaceTest passes
    (on devices with and w/o and updated SM)
Change-Id: I43a5d10c59ae71933db40d932716cd979f802cc5