libbinder: vendor binder has a different header
Right now, on a regular device, libbinder and libbinder.vndk (the
vendor, VNDK variant, of libbinder) are in sync. So, if there is some
error, they will be able to communicate with each other. However, there
is no ABI stability guarantee here. In order to provide an extra layer
of guarantee, we are writing a header to vendor transactions.

This header is not written on system transactions in order to save the
function call, but it may be sane to. Ideally also, it may be sane to
write some hash or similar value instead of a secret code in a specific
case. This would be more flexible.

Bug: 136027762
Test: boot
Change-Id: I664caf6b7082d97e0310e502bf2cd6393e2da9a7