libbinder_ndk Parcelable: no assume non-null write
This is in preparation for supporting @nullable parcelables. Before, for
a parcelable, P::readFromParcel always reads a one. However, this means
that from a 'read nullable parcel' method, we would need to peek the
next int to determine whether or not we want to read the full parcelable
or just return null. In order to avoid this peeking, moving the read of
this int32_t to AParcel_readParcelable. For symmetry,
AParcel_writeParcelable is also added.

Bug: 146172425
Test: atest CtsNdkBinderTestCases
Change-Id: Ie8209979bfd8a2c5b95d129025e5ba5a40c58b9e