fix integer promotion in logging.
Integer promotion rules say that bitfields where all values can be
represented in an int, get promoted to an int. Thus, the computation of
these values in the multiplication is of base integer type.

Compiling this library with debug asserts enabled (-UNDEBUG) currently
fails as the format specifier is expecting long unsigned, and is
provided an int.

Bug: 160324350
Test: Build with -UNDEBUG.
Change-Id: Ia8ab23d2cc6716fe87a36317b1a857981459c207