libbinder: binderParcelTest
For-now-minimal host-supported test to support testing Parcel wire
protocol on host. In the future, this may be expanded to test versions
of the libbinder wire protocol. This is useful because if there is a
minor mistake, when changing the wire protocol, the device will almost
inevitably end up in a bootloop, since binder is used to decrypt
partitions on the device and do other basic early setup.

Bug: 172502290
Test: binderParcelTest
Change-Id: Ie9ac9879bacb8eca3a6c07c8fc4998fe52d8f915