libbinder_ndk: Expose UTF-8 interface descriptor string
We want to be able to expose the original char* interface descriptor
string from AIBinder_Class to NDK clients. Rather than converting the
String16 representation, which presents difficulties regarding
allocation cleanup, we just store an extra copy of the string for now.

Eventually, we would like to transition interface descriptors to UTF-8
and avoid String16 entirely, so this change is a partial step in that
direction and avoids exposing a UTF-16 string through the NDK API.

Bug: 167723746
Test: atest libbinder_ndk_unit_test CtsBinderNdkTestCases
Change-Id: I79d11f26245e229acde753f8c76c53cd0970c8ed