Reland libbinder: vsock support for RPC
This adds support for binder RPC to be used with virtual sockets.

One decision made here - should RpcConnection.h take a sockaddr object,
or not? I've chosen not to expose sockaddr directly because it is a
low-level type (specifically for unix domain sockets, where the path may
not be null-terminated).

Fixes: 182938272
Test: binderRpcTest

Change-Id: Ieb3273f0436502d1624792fd0d7df0270159377e