libbinder: RPC - add DOS TODO
Reading client ID needs to happen on another thread from the main server
in order to avoid a client freezing the server. Though, before moving
the thread tracking to RpcServer, I want to implement reverse
connections (reverse threadpools for handling callbacks) since this is
the feature that originally motivated the threadpool being handled in
RpcSession.

Bug: 185167543
Test: N/A
Change-Id: I1e862ac098159a6efc3c0736050a0b7fe6e7c57e