libbinder_ndk: document headers
libbinder_ndk transactions always include the interface header. This
design decision was originally made in order to be able to rely on this
header for platform features like work source as well as rely on bits
stored there which keep track of API stability and the version of
libbinder being used. In the related bug, we are discussing the
possibility of removing this requirement, but for now documenting it
explicitly.

Bug: 188022639
Test: N/A

Change-Id: I5a83d6eaae9fd7832b322e490ec91f5e4f1005e2