libbinder: run_rpc_tests.sh
There are now a few people working on RPC-related features, and there
are a number of usecases. There have been a couple of cases where
changes broke other tests, but this wasn't realized until later because
presubmit takes a while to run, and running all of the binder tests
locally actually takes a bit of time. Adding this as a quick way to run
RPC tests while they are actively being developed.

Bug: 167966510
Test: each of the following:
  ./run_rpc_tests.sh
  ./run_rpc_tests.sh device
  ./run_rpc_tests.sh host

Change-Id: I5225c6a0ae373546dbef9fa051f50a08f25d7087