libbinder: uptimeMillis returns int64_t!
This was limiting device uptime to ~1 month. Certainly, devices should
update more frequently, but that decision should not be because of
ubsan!

This is a partial copy of the change
Ifa00c32ee184795b03e16b39ac70ff41ebf00331 here.

Bug: 197336441
Test: TH
Merged-In: Ifa00c32ee184795b03e16b39ac70ff41ebf00331
Change-Id: Ib9f3e63e968ac77f0f6ea6447d586db588e9dbe4