[binder_rpc_fuzzer] Depend on shared libutils for the device
Bug: http://b/197965342

binder_rpc_fuzzer depends on libbinder, which already depends on
libutils.so.  When linking binder_rpc_fuzzer with libutils.a, the linker
tries to fetch objects from libutils.a to define undefined symbols in
libbinder.so.  This causes a linker failure with upstream LLD.
Switching to shared libutils.so instead doesn't increase the dependency
closure for binder_rpc_fuzzer.

Test: build with aosp/1809741 which has lld with the above behavior.
Change-Id: I7a3076989eb54669c6a2a07734ef1745033a057e