binderRpcTest: fix CallingId test failure
This test was failing on-device only, and only when certain other tests
were run on it (after ProcessState detects fork). This is an example of
where the fork detection logic is actually being over zealous (we didn't
use kernel binder in the other process, so we actually introduce an
abort). However, this is incorrect usage, and the other process couldn't
actually use kernel binder for this reason. Global state: not even once!

Fixes: 204265622
Test: binderRpcTest (as a whole and also these specific cases, on host
and on device)

Change-Id: Id5bceb596127479e3156bb548ecbe4f60f4f4f0a