service: remove re-implemented functions
We have 'ostream << (const String16&)' and
'IBinder::getInterfaceDescriptor' now, so we don't need implementations
here.

One side-effect of this change is that rather than printing only the
ascii characters from a utf-16 interface type, we will actually convert
the name to utf-8. This would increase the clarity of the output, but
we don't actually expect to benefit from this since interface names
are in utf-8.

Bug: N/A
Test: 'aservice list'
Change-Id: If9bbcf0cdd5be06974d979d9ad7cffb37efd930e