Add servicemanager.microdroid
servicemanager.microdroid is the servicemanager built for microdroid.
The only difference is that it's built with `bootstrap: true` so that it
can start early when APEXes are not activated.

Previously, servicemanager was started after APEXes are activated, but
the introduction of diced made it impossible. Microdroid_manager depends
on diced to select the list of APEXes to be activated. diced in turn
depends on servicemanager.

Bug: 214231981
Test: run microdroid
Change-Id: I6199e1f343d0890f6bf5498e588b92df55d87fe6