libbinder(_ndk): document getService
This API does busy polling, and it returns after a certain number
of tries. There are a few implications related to this, which
most clients (especially in core Android) follow. However, it's
also still pretty easy to introduce errors/races when this API
is used improperly, so clearly discuss how to use it.

Fixes: 218920831
Test: N/A
Change-Id: I1a58aa4517eec8474b01fd94de7241ad3a228b1b