Add #ifdef to avoid calling the system function.
This function is to detect whether CONFIG_ANDROID_BINDER_IPC_32BIT is included in the configuration. It needs to call the system function to start a sub-thread to complete commands such as zcat and grep. However, NuttX currently does not set CONFIG_ANDROID_BINDER_IPC_32BIT, so this detection will cause a lot of unnecessary overhead. Therefore, it is recommended to add #ifdef to disable the relevant detection.

Signed-off-by: cuiziwei <cuiziwei@xiaomi.com>