external/binder:obtain binder latency data.
This program is used to test the performance of Binder IPC. It simulates communication between multiple clients and servers by creating multiple processes and threads, and measures the latency of each transaction.

The main testing process of this program involves the interaction of two transactions in each process, one using FIFO threads for interaction, and the other using other threads for interaction.

Signed-off-by: cuiziwei <cuiziwei@xiaomi.com>