binderLibTest:fix libtest fail and crush error.
After each execution of TEST_F, the child process exits through the destructor function, while other child processes wait for pipe communication. If the read function is interrupted by a sigchild signal at this time, it will cause the pipe to fail to read the content, resulting in a test failure.

Signed-off-by: cuiziwei <cuiziwei@xiaomi.com>