binderLibTest:fix InheritRt test fail.
In TEST_F(BinderLibTest, InheritRt), the priority of the binderLibTest_main program is set to 101. After the setting, it is not restored to 100. This can cause the child thread to fail to exit normally due to its lower priority when the main function exits. Therefore, after the program is executed, the priority is set back to 100.

Signed-off-by: cuiziwei <cuiziwei@xiaomi.com>