Binder: Fix BBinder send to remote memleak
When create a BBinder object, and send it to remote, after process
exit, the remote still hold the BBinder's ref. In Linux OS, the
mem will be freed after process destroyed, while in Nuttx, the mem
will never be freed. So decrease the strong and weak refs when called
ProcessState's destructor.

Signed-off-by: zhangbo56 <zhangbo56@xiaomi.com>